Objects: the devil's four ways, clamped subTypes, and the small guards the sweep listed
Enemy drew random(0, 4) for a direction - random() is inclusive, so the
fifth value folded onto up and the devil walked up twice as often.
Mirror, Teleporter and E_FlipFlop clamp a subType from a foreign level
file in the constructor with a warning, as E_Gate does, instead of
reflecting nothing, teleporting nobody or having no pins while the
tooltip claims a kind. Presets constructs the six electronics parts
after the attribute block and tests p_element for "Damage".

Arrow::turn() wraps dir and shownDir together, so a save carries 0..3
and the animation keeps its direction. Laser's destroyTime stops at the
burst, which fires on the step that reaches 0. The laser's and the light
barrier's beam renderers start their vectors at 0 and always keep the
first two points instead of comparing the second against garbage; the
light barrier's counter advances once per tick, outside the beam loop,
as the laser's does, so its pulse runs at one rate whatever the beam's
length. Lightning's secondary branch walks a Vec2d. Eye tests the enemy
it spawned, Pin::disconnect() writes the input's values directly
(setValue refuses inputs), the TinyXML locals in Elevator and
Electronics start at 0, and E_LightBarrierReceiver no longer saves a
value frameBegin() zeroes before anything reads it.

Dead weight: pin.cpp's EDGY_CONNECTIONS block, which could not have
compiled, and connect()'s unread p_output; e_barrage.h's guard is named
after its file. The flash comment in object.cpp and objects.md says what
FLASH_DECAY = 0.8 does: 25 ticks, half a second. Sprites::add() says
that it reuses the last slot past MAX_SPRITES.

+++ Blocks5/src/lightning.cpp
@@ -153,7 +153,7 @@ Lightning::Branch Lightning::generateSecondaryBranch(const Branch& b,
 
 	// find two consecutive points
 	int start = random(0, static_cast<int>(b.points.size()) - 2);
-	Vec2i pos = b.points[start];
+	Vec2d pos = b.points[start];
 	Vec2d dir = (b.points[start + 1] - b.points[start]).normalize();
 	int length = random(minLength, maxLength);
 	for(int i = 0; i < length; i++)
